travel-agent-ai/backend/requirements.txt
# Core
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.28
alembic==1.13.1
psycopg2-binary==2.9.9

# AI/ML
langgraph==0.0.20
crewai==0.28.4
openai==1.14.2
langchain==0.1.12
langchain-openai==0.0.8

# Utils
pydantic==2.6.4
pydantic-settings==2.2.1
python-dateutil==2.9.0.post0

# Testing
pytest==8.0.2
httpx==0.27.0
pytest-asyncio==0.23.5

# Deployment
gunicorn==21.2.0
uvicorn[standard]==0.27.1